+++ src/pages/index.tsx
@@ -1,13 +1,8 @@
 import { GetServerSideProps, InferGetServerSidePropsType, NextPage } from 'next'
-import { useCallback, useState } from 'react'
 import { CheckBoxGroup } from 'components/CheckBoxGroup'
 import { LineGraph } from 'components/LineGraph'
-import {
-  PrefDetail,
-  PrefDetailResponse,
-  PrefInfo,
-  PrefsResponse,
-} from 'types/prefecture'
+import { usePrefecture } from 'hooks/usePrefecture'
+import { PrefInfo, PrefsResponse } from 'types/prefecture'
 import { getApi } from 'utils/getApi'
 
 type ServerSideProps = {
@@ -16,31 +11,7 @@ type ServerSideProps = {
 type PageProps = InferGetServerSidePropsType<typeof getServerSideProps>
 
 const Home: NextPage<PageProps> = ({ prefs }) => {
-  const [selectedPrefs, setSelectedPrefs] = useState<PrefDetail[]>([])
-
-  const addPref = useCallback(
-    async (prefCode: number) => {
-      const newData = await getApi<PrefDetailResponse>(
-        `https://opendata.resas-portal.go.jp/api/v1/population/composition/perYear?prefCode=${prefCode}`
-      )
-      setSelectedPrefs((postData) => [
-        ...postData,
-        {
-          prefCode,
-          prefName: prefs.find((v) => v.prefCode === prefCode)?.prefName || '',
-          data:
-            newData.result.data.find((v) => v.label === '総人口')?.data || [],
-        },
-      ])
-    },
-    [prefs]
-  )
-
-  const deletePref = useCallback((prefCode: number) => {
-    setSelectedPrefs((postData) =>
-      postData.filter((v) => v.prefCode !== prefCode)
-    )
-  }, [])
+  const { selectedPrefs, addPref, deletePref } = usePrefecture(prefs)
 
   return (
     <>
